MODIFY: Comment 영역 콤포넌트로 분리
