+++ src/pages/namjanghyun/Main/components/ViewComment.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export default ViewComment;
